Fixed compiler warnings in Xcode 4.4.

+++ SuperFoulEgg/UI/include/MenuRectLayer.h
@@ -4,12 +4,12 @@
 #import "CCLayer.h"
 
 @interface MenuRectLayer : CCLayer {
-	int _selectedIndex;
+	NSUInteger _selectedIndex;
 	NSMutableArray *_rectangles;
 }
 
 @property (readonly, retain, nonatomic) NSMutableArray *rectangles;
-@property (readwrite, nonatomic) int selectedIndex;
+@property (readwrite, nonatomic) NSUInteger selectedIndex;
 
 - (void)selectNext;
 - (void)selectPrevious;
